Labyrinth portal redesign: description and remove old code
Labs disable autoexplore and force the player to manually solve a maze,
something not related to the kind of gameplay crawl tries to provide.
There are spoilery aspects to navigating the maze efficiently and to
using teleport in the portal. Once these aspects are known to an
experienced player, they solve the maze with very little thought and
often with great annoyance at having to play this mini-game unrelated to
either tactical combat or strategy. The most interesting aspect of
current labs is the fight with a single powerful monster, the minotaur.

This commit is the first in a series to redesign the portal to be a
gauntlet of arena fights where the player chooses their path. The new
portal emphasizes this decision, giving the player information about the
risk they'll face and rewards they'll get. The portal is similar to
Ziggurats in terms of having progressive rooms with thematic monster
sets, but adds a distinct choice aspect.

Each of the possible paths takes the player through a series of
randomized arena subvaults that have one of two difficulty tiers. Each
arena subvault will having a small group of thematic monsters and a
small amount of loot. All possible paths that could be taken will lead
to the same exit subvault that has boss fight with the minotaur and a
loot pile on the exit stairs.

The number of arena subvaults in each path will be small, usually two
but it could be three or four on some maps. The vaults will will be
accessed via transporters that prevent backtracking by simply walking
out of the vault, and the arenas will not have diggable walls.
Allowing or disallowing teleport is one aspect under consideration for
this portal, but currently teleports are allowed.

Before choosing a path, the player will be able examine all arenas in
each path to see available monsters and loot. At least one early exit
will be provided at the beginning of the portal before any path is
taken.

This commit removes the C++/lua code and vaults that won't be directly
used. Some of the minotaur exit vaults will show up in the new portal
with modifications, but we remove these for now. All that's left in
lab.des is the portal entry vault code needed for save compat and a
dummy lab vault for the portal layout that only places an exit. Many of
the removed float vaults and exit vaults will show up in subsequent
commits, reworked for the new portal.

I'm not yet sure if this new portal will use the name 'Labyrinth', so
I'm leaving lab.des in-place for now instead of moving what's left over
to save compat files. If the portal name does change, it'd be best for
save compat to simply remove any already placed lab entrances.