Lab arena exit subvault loot function
This function places a loot pile on the minotaur's exit with an item
number and composition similar to the old portal. The item count range
is the same the old portal (7 to 16 items with an average of 11.5). The
loot items are 1/3 on-average star_item, which is close to the original
portal's 1/3 chance to have each item be normal class with quality 3 *
depth (star_item is 2 * depth + 5).

The other 2/3 items are a mix of superb_item together with some extra
armour, and some weight for wands and scrolls, since superb_item doesn't
place the latter two. The weights are chosen so that the loot items will
be similar to the custom list used in the old Lab. Gone is the chance
for ammo, and the chance for non-wand misc items and staves are higher,
but in general the differences in chances for each kind of item are not
drastic.

The loot definitions for the old portal were likely not chosen with any
particular aim in mind. Also the new lab arenas will give the player 4-8
items on the path to the minotaur, depending on the map. This does
require defeating nasty monsters, and the old lab gave vault loot along
the way as well. Still, we may want to reconsider this loot composition
in the future.