Lab arena subvault monster sets
This commit defines arena monster sets for two tiers of difficulty. The
sets define one to three related monster sets and define min/max placed
for each. Each set has thematically related monsters that are generally
significantly weaker than the minotaur monster boss. There's an emphasis
on using monsters that might make for an interesting fight for a
character at depths from Lair to D:14, especially for monsters that are
not encountered much in general. Some sets in the same tier are stronger
than others for specific player characters, and some arenas may be too
risky to take on, but this is part of the choice aspect this portal
emphasizes.